Finished the ArchivedPostsViewComponent
1. Finished up the Invoke() method inside of the ArchivedViewComponent
class. This is what calls the GetArchivedPosts method from the dbContext

2. Added a new field caleld "Key" on the Post model. The get method
returns a cleaned up, URL friendly version of the title. This is used
with a new custom route, which is
/posts/year/month/the-clean-title

***HUGE ANNOYANCE***: Had to add the [NotMapped] data annotation to the
Key field. Wrestled with build errors for an hour. This is apparently a
newer issue with newest build of EF.