Finished the AJAX functionality
Added in some basic JS in the site.js file that makes an AJAX call back
to the server. The controller intercepts this request and then checks to
see if it's an AJAX call. If so, sends back a Partial view instead.

End result: I'm only re-rendering the body of the page - just going back
to the server and fetching the 3 posts